fix warnings

+++ backend/PlanetsCall/PlanetsCall/Controllers/Community/OrganisationRolesController.cs
@@ -1,9 +1,9 @@
-﻿using Data.DTO.Community;
+﻿using Core.Exceptions;
+using Data.DTO.Community;
 using Data.Models;
 using Data.Repository.Community;
 using Data.Repository.User;
 using Microsoft.AspNetCore.Mvc;
-using PlanetsCall.Controllers.Exceptions;
 using PlanetsCall.Filters;
 
 namespace PlanetsCall.Controllers.Community;
@@ -23,13 +23,18 @@ public class OrganisationRolesController(
     {
         Users? requestUser = HttpContext.GetRouteValue("requestUser") as Users;
         Organisations? organisation = HttpContext.GetRouteValue("Organisation") as Organisations;
-        if (organisation!.IsPrivate && !organisation.Members.Contains(requestUser))
+        if (organisation!.IsPrivate && organisation.Members != null && requestUser != null && !organisation.Members.Contains(requestUser))
         {
-            return BadRequest(new ErrorResponse(new List<string>() { "Access denied" }, StatusCodes.Status400BadRequest, HttpContext.TraceIdentifier));
+            return BadRequest(new ErrorResponse(["Access denied"], StatusCodes.Status400BadRequest, HttpContext.TraceIdentifier));
         }
 
-        var roles = organisation.Roles.Select(r => new FullRoleDto(r)).ToList();
-        return Ok(roles);
+        if (organisation.Roles != null)
+        {
+            var roles = organisation.Roles.Select(r => new FullRoleDto(r)).ToList();
+            return Ok(roles);
+        }
+
+        return NotFound();
     }
 
     [HttpPost]
@@ -41,14 +46,14 @@ public class OrganisationRolesController(
         
         try
         {
-            organisationsRepository.EnsureUserHasPermission(requestUser!, organisation!.UniqueName, role => role.CanConfigureRoles);
+            organisationsRepository.EnsureUserHasPermission(requestUser!, organisation!.UniqueName, roles => roles.CanConfigureRoles);
             FullRoleDto newRole = organisationsRepository.CreateRole(organisation, role);
 
             return Ok(newRole);
         }
         catch (CodeException e)
         {
-            return StatusCode(e.Code ,new ErrorResponse(new List<string>() { e.Message }, e.Code, HttpContext.TraceIdentifier));
+            return StatusCode(e.Code ,new ErrorResponse([e.Message], e.Code, HttpContext.TraceIdentifier));
         }
     }
 
@@ -62,16 +67,21 @@ public class OrganisationRolesController(
 
         try
         {
-            organisationsRepository.EnsureUserHasPermission(requestUser, organisation.UniqueName,
-                role => role.CanConfigureRoles);
-            FullRoleDto newRoleData = organisationsRepository.UpdateRole(organisation, role, roleId);
+            if (requestUser != null && organisation != null)
+            {
+                organisationsRepository.EnsureUserHasPermission(requestUser, organisation.UniqueName,
+                    roles => roles.CanConfigureRoles);
+                FullRoleDto newRoleData = organisationsRepository.UpdateRole(organisation, role, roleId);
 
-            return Ok(newRoleData);
+                return Ok(newRoleData);
+            }
         }
         catch (CodeException e)
         {
-            return StatusCode(e.Code ,new ErrorResponse(new List<string>() { e.Message }, e.Code, HttpContext.TraceIdentifier));
+            return StatusCode(e.Code ,new ErrorResponse([e.Message], e.Code, HttpContext.TraceIdentifier));
         }
+
+        return BadRequest();
     }
     [HttpDelete]
     [Route("{roleId}/")]
@@ -83,16 +93,20 @@ public class OrganisationRolesController(
 
         try
         {
-            organisationsRepository.EnsureUserHasPermission(requestUser, organisation.UniqueName,
-                role => role.CanConfigureRoles);
-            organisationsRepository.DeleteRole(roleId);
-
-            return NoContent();
+            if (requestUser != null && organisation != null)
+            {
+                organisationsRepository.EnsureUserHasPermission(requestUser, organisation.UniqueName,
+                    role => role.CanConfigureRoles);
+                organisationsRepository.DeleteRole(roleId);
+                return NoContent();
+            }
         }
         catch (CodeException e)
         {
-            return StatusCode(e.Code ,new ErrorResponse(new List<string>() { e.Message }, e.Code, HttpContext.TraceIdentifier));
+            return StatusCode(e.Code ,new ErrorResponse([e.Message], e.Code, HttpContext.TraceIdentifier));
         }
+        
+        return BadRequest();
     }
     [HttpPost]
     [Route("{roleId}/grant/{userId}/")]
@@ -106,17 +120,23 @@ public class OrganisationRolesController(
         {
             Users? user = usersRepository.GetUserById(userId);
             if (user is null) return NotFound();
-            
-            organisationsRepository.EnsureUserHasPermission(requestUser, organisation.UniqueName,
-                role => role.CanConfigureRoles);
-            organisationsRepository.GrantRole(organisation, user, roleId);
 
-            return NoContent();
+
+            if (requestUser != null && organisation != null)
+            {
+                organisationsRepository.EnsureUserHasPermission(requestUser, organisation.UniqueName,
+                    role => role.CanConfigureRoles);
+                organisationsRepository.GrantRole(organisation, user, roleId);
+                
+                return NoContent();
+            }
         }
         catch (CodeException e)
         {
-            return StatusCode(e.Code ,new ErrorResponse(new List<string>() { e.Message }, e.Code, HttpContext.TraceIdentifier));
+            return StatusCode(e.Code ,new ErrorResponse([e.Message], e.Code, HttpContext.TraceIdentifier));
         }
+        
+        return BadRequest();
     }
     
     [HttpPost]
@@ -131,16 +151,21 @@ public class OrganisationRolesController(
         {
             Users? user = usersRepository.GetUserById(userId);
             if (user is null) return NotFound();
-            
-            organisationsRepository.EnsureUserHasPermission(requestUser, organisation.UniqueName,
-                role => role.CanConfigureRoles);
-            organisationsRepository.RevokeRole(organisation, user, roleId);
 
-            return NoContent();
+            if (requestUser != null && organisation != null)
+            {
+                organisationsRepository.EnsureUserHasPermission(requestUser, organisation.UniqueName,
+                    role => role.CanConfigureRoles);
+                organisationsRepository.RevokeRole(organisation, user, roleId);
+
+                return NoContent();
+            }
         }
         catch (CodeException e)
         {
-            return StatusCode(e.Code ,new ErrorResponse(new List<string>() { e.Message }, e.Code, HttpContext.TraceIdentifier));
+            return StatusCode(e.Code ,new ErrorResponse([e.Message], e.Code, HttpContext.TraceIdentifier));
         }
+        
+        return BadRequest();
     }
 }
